lyy-10.18

+++ src/main/java/Command/Goods.java
@@ -4,7 +4,7 @@ public class Goods {
 
     private String name = "cola";
     private String id = "10001";
-    private int price = 10;
+    private double price = 10;
 
     public void buy(){
         System.out.println("Goods [ Name: "+name+", Price: " + price +" ] bought");
